+++ lib/heatmap_calender_direction.dart
@@ -1,1 +1,0 @@
-enum HeatmapCalenderDirection { NOW_TO_TOMORROW, NOW_TO_YESTERDAY }